dnsdist: refactor some common code between doq/doh3
