+++ pdns/dnsdistdist/doq.hh
@@ -37,16 +37,7 @@ struct DownstreamState;
 
 #ifdef HAVE_DNS_OVER_QUIC
 
-/* from rfc9250 section-4.3 */
-enum class DOQ_Error_Codes : uint64_t
-{
-  DOQ_NO_ERROR = 0,
-  DOQ_INTERNAL_ERROR = 1,
-  DOQ_PROTOCOL_ERROR = 2,
-  DOQ_REQUEST_CANCELLED = 3,
-  DOQ_EXCESSIVE_LOAD = 4,
-  DOQ_UNSPECIFIED_ERROR = 5
-};
+#include "doq-common.hh"
 
 struct DOQFrontend
 {
@@ -60,9 +51,8 @@ struct DOQFrontend
   void setup();
 
   std::unique_ptr<DOQServerConfig> d_server_config;
-  TLSConfig d_tlsConfig;
+  dnsdist::doq::QuicheParams d_quicheParams;
   ComboAddress d_local;
-  std::string d_keyLogFile;
 
 #ifdef __linux__
   // On Linux this gives us 128k pending queries (default is 8192 queries),
@@ -71,16 +61,11 @@ struct DOQFrontend
 #else
   uint32_t d_internalPipeBufferSize{0};
 #endif
-  uint64_t d_idleTimeout{5};
-  uint64_t d_maxInFlight{65535};
-  std::string d_ccAlgo{"reno"};
 
   pdns::stat_t d_doqUnsupportedVersionErrors{0}; // Unsupported protocol version errors
   pdns::stat_t d_doqInvalidTokensReceived{0}; // Discarded received tokens
   pdns::stat_t d_validResponses{0}; // Valid responses sent
   pdns::stat_t d_errorResponses{0}; // Empty responses (no backend, drops, invalid queries, etc.)
-
-  static std::map<const string, int> s_available_cc_algorithms;
 };
 
 struct DOQUnit
